# generator/bin/main/akirolabs/tokengenerator/generated/model/Error.kt
package akirolabs.tokengenerator.generated.model

import java.util.Objects
import akirolabs.tokengenerator.generated.model.FaultDefinition
import com.fasterxml.jackson.annotation.JsonProperty
import jakarta.validation.constraints.DecimalMax
import jakarta.validation.constraints.DecimalMin
import jakarta.validation.constraints.Email
import jakarta.validation.constraints.Max
import jakarta.validation.constraints.Min
import jakarta.validation.constraints.NotNull
import jakarta.validation.constraints.Pattern
import jakarta.validation.constraints.Size
import jakarta.validation.Valid
import io.swagger.v3.oas.annotations.media.Schema

/**
* 
    * @param fault 
*/
data class Error(
    
    @field:Valid    
    


    @Schema(example = "null", required = true, description = "")
    @get:JsonProperty("fault", required = true) val fault: FaultDefinition
) {

}