# validator/bin/main/akirolabs/tokenvalidator/generated/model/Token.kt
package akirolabs.tokenvalidator.generated.model

import java.util.Objects
import com.fasterxml.jackson.annotation.JsonProperty
import jakarta.validation.constraints.DecimalMax
import jakarta.validation.constraints.DecimalMin
import jakarta.validation.constraints.Email
import jakarta.validation.constraints.Max
import jakarta.validation.constraints.Min
import jakarta.validation.constraints.NotNull
import jakarta.validation.constraints.Pattern
import jakarta.validation.constraints.Size
import jakarta.validation.Valid
import io.swagger.v3.oas.annotations.media.Schema

/**
* 
    * @param token 
*/
data class Token(
    
    


    @Schema(example = "1234-5678-9012-3456", description = "")
    @get:JsonProperty("token") val token: kotlin.String? = null
) {

}